Rename test_be module as test_bel
